Fix integration-suite put tests for writing relationships and sets
The getGun() function appears to instantiate a new Gun instance each time it is called.
It is called in a beforeEach() block prior to each test and sets up the gun
variable with the instance.
However, in certain tests, getGun() is called separately midway through the test.
This looks to cause the test to try to assert data in a
different Gun instance than the data was written to prior in the test,
leading to a test failure.
Replacing the getGun() call in these tests with the gun variable instance
that was setup in the beforeEach block appears to fix the issue and make
the tests pass.